fix subscribe tests
