+++ examples/protocols/src/main.rs
@@ -1,6 +1,6 @@
 use base64ct::{Base64UrlUnpadded, Encoding};
 use dwn_test::keystore::ALICE_DID;
-use dwn_test::store::ProviderImpl;
+use dwn_test::provider::ProviderImpl;
 use serde_json::json;
 use vercre_dwn::endpoint;
 use vercre_dwn::protocols::Query;
