+++ tests/messages_query.rs
@@ -4,7 +4,7 @@
 //! messages and subsequently query for them.
 
 use dwn_test::keystore::ALICE_DID;
-use dwn_test::store::ProviderImpl;
+use dwn_test::provider::ProviderImpl;
 use http::StatusCode;
 // use insta::assert_yaml_snapshot as assert_snapshot;
 use serde_json::json;
